v0.2 Alpha
New Update! v0.2 Alpha
Features:
- Prices will now have a constant gap between them, instead of a somewhat arbitrary gap based on the icon width.
- Partially rewritten price display code
- Cleaned *some* code
- Somewhat deprecating commodities, the code still exists, but will need to be uncommented. I don't like using the yahoo API and I am trying to find a better way to get stock prices as well. Future updates may or may not include support for commodities.

Features that are probably coming in the next update:
- On/Off time, stops displaying after a certain time, and back on at a certain time (because it hurts my eyes when I am trying to sleep lol)
- Maybe some code cleaning/ refactoring of variables to make more sense.

Features that are planned:
- Fix the GET requests so they actually happen in the background and don't freeze the display
- Move some things to a config file
- Some more types of displays (Bitcoin block height/ difficulty, etc.)

Please do not expect regular updates, my final year in high school is pretty tough, so unfortunately most of the time I've got more important things to be thinking about than this. Once again I would highly encourage anyone to help to contribute to this project, when we work as a team magical things happen.

Until next time (or update),
Puffball (Joel)